CG-133 Refactor: app/custom-components/ 내 폴더들 정리
- custom-components/basics/ 로 모음
- 이에따라, index.js 도 수정보완함
