+++ app/custom-components/buttons/come-home-go-to-switch-button/come-home-go-to-switch-button.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { ViewStyle, View } from "react-native"
 import { color, WIDTH, HEIGHT, palette } from "~/app/theme"
 import { DISABLED, GIVER_CASUAL_NAVY } from "~/app/theme/palette"
-import { PreBol12 } from "../../custom-texts/custom-texts"
+import { PreBol12 } from "../../basics/custom-texts/custom-texts"
 import { PressableButton } from "../pressable-button/pressable-button"
 
 const ROOT: ViewStyle = {
