+++ app/custom-components/buttons/conditional-button/conditional-button.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { Pressable, Text, PressableProps } from "react-native"
 import { palette } from "~/app/theme"
-import { PreBol16 } from "../../custom-texts/custom-texts"
+import { PreBol16 } from "../../basics/custom-texts/custom-texts"
 import { PressableButton } from "../pressable-button/pressable-button"
 import { styles } from "./styles"
