+++ app/custom-components/buttons/service-choice-button/service-choice-button.tsx
@@ -2,7 +2,7 @@ import { View, Text, Pressable, Image } from "react-native"
 import React from "react"
 import { styles } from "./styles"
 import { ChoiceButtonProps } from "./service-choice-button.props"
-import { PreBol16, PreReg12 } from "../../custom-texts/custom-texts"
+import { PreBol16, PreReg12 } from "../../basics/custom-texts/custom-texts"
 import { SHADOW_1 } from "~/app/theme"
 import { BODY, SUB_HEAD_LINE } from "~/app/theme/palette"
 import IMAGES from "~/assets/images"
