+++ app/custom-components/caregiver-certificate/caregiver-certificate.tsx
@@ -2,7 +2,7 @@ import { View, Image, Pressable } from "react-native"
 import React from "react"
 import { styles } from "./styles"
 import { HEIGHT, WIDTH } from "~/app/theme"
-import { PreReg12, PreReg16 } from "../custom-texts/custom-texts"
+import { PreReg12, PreReg16 } from "../basics/custom-texts/custom-texts"
 import { SUB_HEAD_LINE } from "~/app/theme/palette"
 import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
