+++ app/custom-components/caregiver-name-star-review/caregiver-name-star-review.tsx
@@ -2,7 +2,7 @@ import { View, Text, Pressable, Image } from "react-native"
 import React from "react"
 import { styles } from "./styles"
 import { WIDTH, HEIGHT } from "~/app/theme"
-import { PreBol14, PreMed16 } from "../custom-texts/custom-texts"
+import { PreBol14, PreMed16 } from "../basics/custom-texts/custom-texts"
 import { DBG, GIVER_CASUAL_NAVY, HEAD_LINE, SUB_HEAD_LINE } from "~/app/theme/palette"
 import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
