+++ app/custom-components/caregiver-service/caregiver-service.tsx
@@ -2,7 +2,7 @@ import { View, Image } from "react-native"
 import React from "react"
 import { styles } from "./styles"
 import { HEIGHT, WIDTH } from "~/app/theme"
-import { PreMed16 } from "../custom-texts/custom-texts"
+import { PreMed16 } from "../basics/custom-texts/custom-texts"
 import { BODY } from "~/app/theme/palette"
 import { Row } from "../basics/row/row"
 import IMAGES from "~/assets/images"
