+++ app/custom-components/dropdown-boxes/pet-info-dropdown-box/pet-info-dropdown-box.tsx
@@ -4,7 +4,7 @@ import { FlatList } from "react-native-gesture-handler"
 import { PetProfileCard } from "../../pet-profile-card/pet-profile-card"
 import { styles } from "./styles"
 import { Row } from "../../basics/row/row"
-import { PopSem14 } from "../../custom-texts/custom-texts"
+import { PopSem14 } from "../../basics/custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { PetInfoDropdownBoxProps } from "./pet-info-dropdown-box.props"
 import { DivisionLine } from "../../division-line/division-line"
