+++ app/custom-components/dropdown-boxes/select-option-button/select-option-dropdown-box.tsx
@@ -1,7 +1,7 @@
 import { View, Image, Pressable } from "react-native"
 import React from "react"
 import { Row } from "../../basics/row/row"
-import { PreReg12 } from "../../custom-texts/custom-texts"
+import { PreReg12 } from "../../basics/custom-texts/custom-texts"
 import { FlatList } from "react-native-gesture-handler"
 
 export const SelectOptionDropdownBox = (props) => {
