+++ app/custom-components/filter-header/filter-header.tsx
@@ -1,7 +1,7 @@
 import { View, Pressable, Image } from "react-native"
 import React from "react"
 import { Row } from "../basics/row/row"
-import { PopSem16, PreBol16, PreReg12 } from "../custom-texts/custom-texts"
+import { PopSem16, PreBol16, PreReg12 } from "../basics/custom-texts/custom-texts"
 import { HEIGHT, WIDTH } from "~/app/theme"
 import IMAGES from "~/assets/images"
 import { FilterHeaderProps } from "./filter-header.props"
