+++ app/custom-components/hired-times-and-pet-years/hired-times-and-pet-years.tsx
@@ -1,6 +1,6 @@
 import { View, Text, StyleProp, FlexStyle } from "react-native"
 import React from "react"
-import { PreMed14, PreBol16 } from "../custom-texts/custom-texts"
+import { PreMed14, PreBol16 } from "../basics/custom-texts/custom-texts"
 import { BODY, LIGHT_LINE, STRONG_LINE } from "~/app/theme/palette"
 import { styles } from "./styles"
 import { HEIGHT } from "~/app/theme"
