+++ app/custom-components/in-progress-booking/in-progress-booking-profile/in-progress-booking-profile.tsx
@@ -2,7 +2,7 @@ import { View, Text, Image } from "react-native"
 import React from "react"
 import { styles } from "./styles"
 import { Row } from "../../basics/row/row"
-import { PreReg12, PreReg14 } from "../../custom-texts/custom-texts"
+import { PreReg12, PreReg14 } from "../../basics/custom-texts/custom-texts"
 import { WIDTH, palette, HEIGHT } from "~/app/theme"
 import CaregiverTypeButton from "../../buttons/caregiver-type-button/caregiver-type-button"
 import RatingReviewBox from "../../rating-review-box/rating-review-box"
