+++ app/custom-components/in-progress-booking/reserve-date-box/reserve-date-box.tsx
@@ -1,7 +1,7 @@
 import { View, Text, Image } from "react-native"
 import React from "react"
 import { Row } from "../../basics/row/row"
-import { PreBol14, PreReg12 } from "../../custom-texts/custom-texts"
+import { PreBol14, PreReg12 } from "../../basics/custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { styles } from "./styles"
 import { HEIGHT } from "~/app/theme"
