+++ app/custom-components/index.ts
@@ -1,5 +1,8 @@
-//* texts
-export * from "./custom-texts/custom-texts"
+//* basics
+export * from "./basics/custom-texts/custom-texts"
+export * from "./basics/row/row"
+export * from "./basics/row-rounded-box/row-rounded-box"
+export * from "./basics/view-component/view-component"
 
 //* buttons
 export * from "./buttons/caregiver-type-button/caregiver-type-button"
@@ -10,19 +13,16 @@ export * from "./buttons/row-rounded-button/row-rounded-button"
 export * from "./buttons/public-private-switch-button/public-private-switch-button"
 export * from "./buttons/make-booking-button/make-booking-button"
 export * from "./buttons/sitter-profile-button/sitter-profile-button"
-
-//* headers
-export * from "./service-type-indicator-header/service-type-indicator-header"
-export * from "./screen-headers/goback-and-title-header/goback-and-title-header"
-export * from "./screen-headers/home-screen-header/home-screen-header"
-export * from "./screen-headers/all-comments-screen-header/all-comments-screen-header"
-
-//* others
-export * from "./basics/view-component/view-component"
 export * from "./buttons/service-choice-button/service-choice-button"
 export * from "./buttons/sitter-profile-button/sitter-profile-button"
-export * from "./basics/row/row"
-export * from "./basics/row-rounded-box/row-rounded-box"
+
+//* screen-headers
+export * from "./screen-headers/all-comments-screen-header/all-comments-screen-header"
+export * from "./screen-headers/goback-and-title-header/goback-and-title-header"
+export * from "./screen-headers/home-screen-header/home-screen-header"
+export * from "./screen-headers/writing-comment-screen-header/writing-comment-screen-header"
+
+//* others
 export * from "./division-line/division-line"
 export * from "./dots-indicator/dots-indicator"
 export * from "./selected-pet-card/selected-pet-card"
@@ -32,12 +32,7 @@ export * from "./blue-checkbox/blue-checkbox"
 export * from "./full-width-size-images-box-with-indicator/full-width-size-images-box-with-indicator"
 export * from "./hired-times-and-pet-years/hired-times-and-pet-years"
 export * from "./comment/comment"
-
-//* headers
-export * from "./service-type-indicator-header/service-type-indicator-header"
-export * from "./screen-headers/goback-and-title-header/goback-and-title-header"
-export * from "./screen-headers/home-screen-header/home-screen-header"
-export * from "./screen-headers/writing-comment-screen-header/writing-comment-screen-header"
 export * from "./caregiver-certificate/caregiver-certificate"
 export * from "./caregiver-service/caregiver-service"
 export * from "./caregiver-name-star-review/caregiver-name-star-review"
+export * from "./service-type-indicator-header/service-type-indicator-header"
