+++ app/custom-components/screen-headers/all-comments-screen-header/all-comments-screen-header.tsx
@@ -1,7 +1,7 @@
 import { View, Image, Pressable } from "react-native"
 import React from "react"
 import { WIDTH } from "~/app/theme"
-import { PreMed18 } from "../../custom-texts/custom-texts"
+import { PreMed18 } from "../../basics/custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { styles } from "./styles"
 import { navigate } from "~/app/navigators"
