+++ app/custom-components/screen-headers/goback-and-title-header/goback-and-title-header.tsx
@@ -1,7 +1,7 @@
 import { View, Image, Pressable } from "react-native"
 import React from "react"
 import { WIDTH } from "~/app/theme"
-import { PreMed18 } from "../../custom-texts/custom-texts"
+import { PreMed18 } from "../../basics/custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { styles } from "./styles"
 import { HEADER_ROOT } from "../common-styles"
