+++ app/custom-components/screen-headers/writing-comment-screen-header/writing-comment-screen-header.tsx
@@ -1,7 +1,7 @@
 import { View, Image, Pressable } from "react-native"
 import React from "react"
 import { WIDTH } from "~/app/theme"
-import { PreMed20, PreBol16 } from "../../custom-texts/custom-texts"
+import { PreMed20, PreBol16 } from "../../basics/custom-texts/custom-texts"
 import IMAGES from "~/assets/images"
 import { styles } from "./styles"
 import { DISABLED, GIVER_CASUAL_NAVY } from "~/app/theme/palette"
