+++ app/custom-components/selected-pet-card/selected-pet-card.tsx
@@ -1,7 +1,7 @@
 import { View, Text, Pressable, Image } from "react-native"
 import React from "react"
 import { styles } from "./styles"
-import { PreBol16, PreReg12, PreReg14 } from "../custom-texts/custom-texts"
+import { PreBol16, PreReg12, PreReg14 } from "../basics/custom-texts/custom-texts"
 import { BODY, DBG, HEAD_LINE, LBG, SUB_HEAD_LINE } from "~/app/theme/palette"
 import IMAGES from "~/assets/images"
 import { DivisionLine } from "../division-line/division-line"
